# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemini Surgicals API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  shortDescription: string;
  price: number;
  discountPrice?: number | null;
  categoryId?: number | null;
  categoryName?: string | null;
  imageUrl?: string | null;
  stock: number;
  inStock: boolean;
  weeklyOffer: boolean;
  featured: boolean;
  specifications?: string | null;
  deliveryInfo?: string | null;
  createdAt: string;
}

export interface ProductsResponse {
  products: Product[];
  total: number;
  limit: number;
  offset: number;
}

export interface CreateProductInput {
  name: string;
  description: string;
  shortDescription: string;
  price: number;
  discountPrice?: number | null;
  categoryId?: number | null;
  imageUrl?: string | null;
  stock: number;
  weeklyOffer?: boolean;
  featured?: boolean;
  specifications?: string | null;
  deliveryInfo?: string | null;
}

export interface UpdateProductInput {
  name?: string;
  description?: string;
  shortDescription?: string;
  price?: number;
  discountPrice?: number | null;
  categoryId?: number | null;
  imageUrl?: string | null;
  stock?: number;
  weeklyOffer?: boolean;
  featured?: boolean;
  specifications?: string | null;
  deliveryInfo?: string | null;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  imageUrl?: string | null;
  productCount: number;
}

export interface CreateCategoryInput {
  name: string;
  description?: string | null;
  imageUrl?: string | null;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  price: number;
  subtotal: number;
}

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  mpesa: "mpesa",
  cash_on_delivery: "cash_on_delivery",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  location: string;
  paymentMethod: OrderPaymentMethod;
  status: OrderStatus;
  items: OrderItem[];
  totalAmount: number;
  notes?: string | null;
  createdAt: string;
}

export type CreateOrderInputPaymentMethod =
  (typeof CreateOrderInputPaymentMethod)[keyof typeof CreateOrderInputPaymentMethod];

export const CreateOrderInputPaymentMethod = {
  mpesa: "mpesa",
  cash_on_delivery: "cash_on_delivery",
} as const;

export type CreateOrderInputItemsItem = {
  productId: number;
  quantity: number;
};

export interface CreateOrderInput {
  customerName: string;
  phone: string;
  location: string;
  paymentMethod: CreateOrderInputPaymentMethod;
  items: CreateOrderInputItemsItem[];
  notes?: string | null;
}

export type UpdateOrderStatusInputStatus =
  (typeof UpdateOrderStatusInputStatus)[keyof typeof UpdateOrderStatusInputStatus];

export const UpdateOrderStatusInputStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusInput {
  status: UpdateOrderStatusInputStatus;
}

export interface AdminStats {
  totalProducts: number;
  totalOrders: number;
  pendingOrders: number;
  totalRevenue: number;
  lowStockProducts: number;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  inStock?: boolean;
  weeklyOffer?: boolean;
  sort?: ListProductsSort;
  limit?: number;
  offset?: number;
};

export type ListProductsSort =
  (typeof ListProductsSort)[keyof typeof ListProductsSort];

export const ListProductsSort = {
  price_asc: "price_asc",
  price_desc: "price_desc",
  popular: "popular",
  newest: "newest",
} as const;