# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gemini Surgicals API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: cart
    description: Cart operations
  - name: orders
    description: Order operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: weeklyOffer
          in: query
          schema:
            type: boolean
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, popular, newest]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/featured:
    get:
      operationId: listFeaturedProducts
      tags: [products]
      summary: Get featured/offer products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [admin]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /orders:
    get:
      operationId: listOrders
      tags: [admin]
      summary: List all orders (admin)
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        price:
          type: number
        discountPrice:
          type: number
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        stock:
          type: integer
        inStock:
          type: boolean
        weeklyOffer:
          type: boolean
        featured:
          type: boolean
        specifications:
          type: string
          nullable: true
        deliveryInfo:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - description
        - shortDescription
        - price
        - stock
        - inStock
        - weeklyOffer
        - featured
        - createdAt

    ProductsResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [products, total, limit, offset]

    CreateProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        price:
          type: number
        discountPrice:
          type: number
          nullable: true
        categoryId:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        stock:
          type: integer
        weeklyOffer:
          type: boolean
        featured:
          type: boolean
        specifications:
          type: string
          nullable: true
        deliveryInfo:
          type: string
          nullable: true
      required: [name, description, shortDescription, price, stock]

    UpdateProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        price:
          type: number
        discountPrice:
          type: number
          nullable: true
        categoryId:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        stock:
          type: integer
        weeklyOffer:
          type: boolean
        featured:
          type: boolean
        specifications:
          type: string
          nullable: true
        deliveryInfo:
          type: string
          nullable: true

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        productCount:
          type: integer
      required: [id, name, slug, productCount]

    CreateCategoryInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required: [name]

    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
      required: [productId, productName, quantity, price, subtotal]

    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        phone:
          type: string
        location:
          type: string
        paymentMethod:
          type: string
          enum: [mpesa, cash_on_delivery]
        status:
          type: string
          enum: [pending, confirmed, processing, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - customerName
        - phone
        - location
        - paymentMethod
        - status
        - items
        - totalAmount
        - createdAt

    CreateOrderInput:
      type: object
      properties:
        customerName:
          type: string
        phone:
          type: string
        location:
          type: string
        paymentMethod:
          type: string
          enum: [mpesa, cash_on_delivery]
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
            required: [productId, quantity]
        notes:
          type: string
          nullable: true
      required: [customerName, phone, location, paymentMethod, items]

    UpdateOrderStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, delivered, cancelled]
      required: [status]

    AdminStats:
      type: object
      properties:
        totalProducts:
          type: integer
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        totalRevenue:
          type: number
        lowStockProducts:
          type: integer
      required:
        - totalProducts
        - totalOrders
        - pendingOrders
        - totalRevenue
        - lowStockProducts