# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemini Surgicals API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  location: string;
  paymentMethod: OrderPaymentMethod;
  status: OrderStatus;
  items: OrderItem[];
  totalAmount: number;
  notes?: string | null;
  createdAt: string;
}