# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemini Surgicals API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  shortDescription: string;
  price: number;
  discountPrice?: number | null;
  categoryId?: number | null;
  categoryName?: string | null;
  imageUrl?: string | null;
  stock: number;
  inStock: boolean;
  weeklyOffer: boolean;
  featured: boolean;
  specifications?: string | null;
  deliveryInfo?: string | null;
  createdAt: string;
}