# lib/api-zod/src/generated/types/updateOrderStatusInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemini Surgicals API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateOrderStatusInputStatus } from "./updateOrderStatusInputStatus";

export interface UpdateOrderStatusInput {
  status: UpdateOrderStatusInputStatus;
}